requirements.txt
beautifulsoup4
Confluence-py
requests